Add base class for PersistentTopic and NonPersistentTopic
